chg: usr: storing ruleset activation is optionnal

+++ ruleskit/rule.py
@@ -43,7 +43,8 @@ class Rule(ABC):
     def del_activation(self):
         """Deletes the activation vector's data, but not the object itself, so any computed attribute will remain
         available"""
-        self._activation.delete()
+        if self._activation is not None:
+            self._activation.delete()
 
     @property
     def activation_available(self):
